requirements.txt
# Core dependencies
transformers>=4.30.0
torch>=2.0.0
huggingface_hub>=0.16.0

# Security and signing
cosign-python>=0.1.0
garak>=0.9.0

# Guardrails
guardrails-ai>=0.4.0
nemoguardrails>=0.7.0

# API framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.0.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-instrumentation-fastapi>=0.39b0

# Utilities
pyyaml>=6.0
requests>=2.31.0
aiofiles>=23.0.0
python-multipart>=0.0.6
